+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - Add command kidney\_core list tickets
 - add support in add command for multiple repos
 - Add constants.dart and change master folder to .master
+- Add tests for creation of .organizations
+- Add test: logs error when primary and all fallback organization clones fail
 
 ### Changed